+++ src/main/java/br/com/combinacaotaxonomias/helper/CategoriaHelper.java
@@ -12,7 +12,7 @@ public class CategoriaHelper {
 		return categoria;
 	}
 	
-	public  static CategoriaTO toCategoriaTO(Taxonomia categoria, Integer idPai, Integer idPlataforma) {	
+	public  static CategoriaTO toCategoriaTO(Taxonomia categoria, Integer idPai, Long idPlataforma) {	
 		CategoriaTO categoriaTO = new CategoriaTO(categoria.getId().toString(), categoria.getNome(), idPai.toString(), idPlataforma.toString());
 		
 		return categoriaTO;
